clean up transforms/core
