+++ pennylane/transforms/core/transform.py
@@ -14,49 +14,21 @@
 """
 This module defines the data structure that encapsulates a quantum transform.
 """
+
 from __future__ import annotations
 
 import os
 import warnings
-from collections.abc import Callable, Sequence
+from collections.abc import Callable
 from copy import copy
-from functools import lru_cache, partial, singledispatch, update_wrapper, wraps
+from functools import partial, singledispatch, update_wrapper
 
-from pennylane import capture, math
-from pennylane.capture import autograph
-from pennylane.exceptions import PennyLaneDeprecationWarning, TransformError
-from pennylane.measurements import MeasurementProcess
-from pennylane.operation import Operator
-from pennylane.pytrees import flatten
-from pennylane.queuing import AnnotatedQueue, QueuingManager, apply
-from pennylane.tape import QuantumScript
-from pennylane.typing import ResultBatch
+from pennylane import capture
+from pennylane.exceptions import TransformError
 
-
-@lru_cache
-def _create_transform_primitive():
-    try:
-        # pylint: disable=import-outside-toplevel
-        from pennylane.capture.custom_primitives import QmlPrimitive
-    except ImportError:
-        return None
-
-    transform_prim = QmlPrimitive("transform")
-    transform_prim.multiple_results = True
-    transform_prim.prim_type = "transform"
-
-    # pylint: disable=too-many-arguments, disable=unused-argument
-    @transform_prim.def_impl
-    def _impl(*all_args, inner_jaxpr, args_slice, consts_slice, **_):
-        args = all_args[slice(*args_slice)]
-        consts = all_args[slice(*consts_slice)]
-        return capture.eval_jaxpr(inner_jaxpr, consts, *args)
-
-    @transform_prim.def_abstract_eval
-    def _abstract_eval(*_, inner_jaxpr, **__):
-        return [out.aval for out in inner_jaxpr.outvars]
-
-    return transform_prim
+from ._dispatch import generic_apply_transform, specific_apply_transform
+from .bound_transform import BoundTransform
+from .compile_pipeline import CompilePipeline
 
 
 def _create_plxpr_fallback_transform(tape_transform):
@@ -99,31 +71,6 @@ def _create_plxpr_fallback_transform(tape_transform):
         return jax.make_jaxpr(wrapper, abstracted_axes=abstracted_axes)(*abstract_shapes, *args)
 
     return plxpr_fallback_transform
-
-
-def specific_apply_transform(transform, obj, *targs, **tkwargs):
-    """The default behavior for Transform._apply_transform. By default, it dispatches to the
-    generic registration."""
-    return transform.generic_apply_transform(obj, *targs, **tkwargs)
-
-
-@singledispatch
-def generic_apply_transform(obj, transform, *targs, **tkwargs):
-    """Apply an generic transform to a specific type of object. A singledispatch function
-    used by ``TransformDipsatcher.generic_apply_transform``, but with a different order of arguments
-    to allow is to be used by singledispatch.
-
-    When called with an object that is not a valid dispatch target (e.g., not a QNode, tape, etc.),
-    this returns a BoundTransform with the supplied args and kwargs. This enables patterns like:
-
-        decompose(gate_set=gate_set) + merge_rotations(1e-6)
-
-    where transforms are called with just configuration parameters and combined into a CompilePipeline
-    """
-    # If the first argument is not a valid dispatch target, return a BoundTransform
-    # with the first argument and any additional args/kwargs stored as transform parameters.
-    targs, tkwargs = transform.setup_inputs(obj, *targs, **tkwargs)
-    return BoundTransform(transform, args=targs, kwargs=tkwargs)
 
 
 # pragma: no cover
@@ -737,9 +684,6 @@ class Transform:  # pylint: disable=too-many-instance-attributes
             )
 
         if self.expand_transform:
-            # pylint: disable=import-outside-toplevel
-            from .compile_pipeline import CompilePipeline
-
             return CompilePipeline(self, other)
 
         # Convert this transform to a BoundTransform (no args/kwargs) and delegate
@@ -748,9 +692,6 @@ class Transform:  # pylint: disable=too-many-instance-attributes
     def __mul__(self, n):
         """Multiply by an integer to create a compile pipeline with this transform repeated."""
         if self.expand_transform:
-            # pylint: disable=import-outside-toplevel
-            from .compile_pipeline import CompilePipeline
-
             return CompilePipeline(self) * n
 
         # Convert to container (no args/kwargs) and delegate
@@ -843,445 +784,4 @@ class Transform:  # pylint: disable=too-many-instance-attributes
         return qnode
 
 
-class BoundTransform:  # pylint: disable=too-many-instance-attributes
-    """A transform with bound inputs.
-
-    Args:
-        transform: Any transform.
-        args (Sequence[Any]): The positional arguments to use with the transform.
-        kwargs (Dict | None): The keyword arguments for use with the transform.
-
-    Keyword Args:
-        use_argnum (bool): An advanced option used in conjunction with calculating
-            classical cotransforms of jax workflows.
-
-    .. seealso:: :func:`~.pennylane.transform`
-
-    >>> bound_t = BoundTransform(qml.transforms.merge_rotations, (), {"atol": 1e-4})
-    >>> bound_t
-    <merge_rotations(atol=0.0001)>
-
-    The class can also be created by directly calling the transform with its inputs:
-
-    >>> qml.transforms.merge_rotations(atol=1e-4)
-    <merge_rotations(atol=0.0001)>
-
-    These objects can now directly applied to anything individual transforms can apply to:
-
-    .. code-block:: python
-
-        @bound_t
-        @qml.qnode(qml.device('null.qubit', wires=2))
-        def c(x):
-            qml.RX(x, 0)
-            qml.RX(-x + 1e-6, 0)
-            qml.RY(x, 1)
-            qml.RY(-x + 1e-2, 1)
-            return qml.probs(wires=(0,1))
-
-    If we draw this circuit, we can see that the ``merge_rotations`` transforms was applied with a
-    tolerance of ``1e-4``.  The ``RX`` gates sufficiently close to zero disappear, while the ``RY`` gates
-    that are further from zero remain.
-
-    >>> print(qml.draw(c)(1.0))
-    0: ───────────┤ ╭Probs
-    1: ──RY(0.01)─┤ ╰Probs
-
-    Repeated versions of the bound transform can be created with multiplication:
-
-    >>> print(bound_t * 3)
-    CompilePipeline(
-      [1] merge_rotations(atol=0.0001),
-      [2] merge_rotations(atol=0.0001),
-      [3] merge_rotations(atol=0.0001)
-    )
-
-    And it can be used in conjunction with both individual transforms, bound transforms, and
-    compile pipelines.
-
-    >>> print(bound_t + qml.transforms.cancel_inverses)
-    CompilePipeline(
-      [1] merge_rotations(atol=0.0001),
-      [2] cancel_inverses()
-    )
-    >>> print(bound_t + qml.transforms.cancel_inverses + bound_t)
-    CompilePipeline(
-      [1] merge_rotations(atol=0.0001),
-      [2] cancel_inverses(),
-      [3] merge_rotations(atol=0.0001)
-    )
-
-    """
-
-    def __hash__(self):
-        hashable_dict = tuple((key, value) for key, value in self.kwargs.items())
-        return hash((self.tape_transform, self.pass_name, self.args, hashable_dict))
-
-    def __init__(
-        self,
-        transform: Transform,
-        args: tuple | list = (),
-        kwargs: None | dict = None,
-        *,
-        use_argnum: bool = False,
-        **transform_config,
-    ):
-        if not isinstance(transform, Transform):
-            transform = Transform(transform, **transform_config)
-        elif transform_config:
-            raise ValueError(
-                f"transform_config kwargs {transform_config} cannot be passed if a transform is provided."
-            )
-        self._transform = transform
-        self._args = tuple(args)
-        self._kwargs = kwargs or {}
-        self._use_argnum = use_argnum
-
-    def __repr__(self):
-        name = self.tape_transform.__name__ if self.tape_transform else self.pass_name
-        arg_str = ", ".join(repr(a) for a in self._args) if self._args else ""
-        kwarg_str = (
-            ", ".join(f"{key}={value}" for key, value in self._kwargs.items())
-            if self._kwargs
-            else ""
-        )
-        if arg_str and kwarg_str:
-            total_str = ", ".join([arg_str, kwarg_str])
-        elif arg_str:
-            total_str = arg_str
-        else:
-            total_str = kwarg_str
-        return f"<{name}({total_str})>"
-
-    def __call__(self, obj):
-        return self._transform(obj, *self.args, **self.kwargs)
-
-    def __iter__(self):
-        return iter(
-            (
-                self._transform.tape_transform,
-                self._args,
-                self._kwargs,
-                self._transform.classical_cotransform,
-                self._transform.plxpr_transform,
-                self._transform.is_informative,
-                self._transform.is_final_transform,
-            )
-        )
-
-    def __eq__(self, other: object) -> bool:
-        if not isinstance(other, BoundTransform):
-            return False
-        return (
-            self.args == other.args
-            and self.tape_transform == other.tape_transform
-            and self.pass_name == other.pass_name
-            and self.kwargs == other.kwargs
-            and self.classical_cotransform == other.classical_cotransform
-            and self.is_informative == other.is_informative
-            and self.is_final_transform == other.is_final_transform
-        )
-
-    @property
-    def tape_transform(self) -> Callable | None:
-        """The raw tape transform definition for the transform."""
-        return self._transform.tape_transform
-
-    @property
-    def transform(self) -> Callable | None:
-        """The raw tape transform definition of the transform.
-
-        .. warning::
-            This property is deprecated and will be removed in v0.46.
-            Please use :attr:`~.BoundTransform.tape_transform` instead.
-
-        """
-        warnings.warn(
-            "The 'BoundTransform.transform' property is deprecated and will be removed in v0.46. "
-            "Please use 'BoundTransform.tape_transform' instead.",
-            PennyLaneDeprecationWarning,
-            stacklevel=2,
-        )
-        return self.tape_transform
-
-    @property
-    def expand_transform(self) -> BoundTransform | None:
-        """The expand_transform associated with this transform."""
-        if not self._transform.expand_transform:
-            return None
-        return BoundTransform(
-            self._transform.expand_transform,
-            args=self.args,
-            kwargs=self.kwargs,
-            use_argnum=self._transform._use_argnum_in_expand,  # pylint:disable=protected-access
-        )
-
-    @property
-    def pass_name(self) -> None | str:
-        """The name of the corresponding Catalyst pass, if it exists."""
-        return self._transform.pass_name
-
-    @property
-    def args(self) -> tuple:
-        """The stored quantum transform's ``args``."""
-        return self._args
-
-    @property
-    def kwargs(self) -> dict:
-        """The stored quantum transform's ``kwargs``."""
-        return self._kwargs
-
-    @property
-    def classical_cotransform(self) -> None | Callable:
-        """The stored quantum transform's classical co-transform."""
-        return self._transform.classical_cotransform
-
-    @property
-    def plxpr_transform(self) -> None | Callable:
-        """The stored quantum transform's PLxPR transform.
-
-        **UNMAINTAINED AND EXPERIMENTAL**
-        """
-        return self._transform.plxpr_transform
-
-    @property
-    def is_informative(self) -> bool:
-        """Whether or not a transform is informative. If true the transform is queued at the end
-        of the transform program and the tapes or qnode aren't executed.
-
-        This property is rare, but used by such transforms as ``qml.transforms.commutation_dag``.
-        """
-        return self._transform.is_informative
-
-    @property
-    def is_final_transform(self) -> bool:
-        """Whether or not the transform must be the last one to be executed
-        in a ``CompilePipeline``.
-
-        This property is ``True`` for most gradient transforms.
-        """
-        return self._transform.is_final_transform
-
-    def __add__(self, other):
-        """Add two transforms to create a CompilePipeline."""
-
-        if not isinstance(other, (Transform, BoundTransform)):
-            return NotImplemented
-
-        # Import here to avoid circular import
-        # pylint: disable=import-outside-toplevel
-        from .compile_pipeline import CompilePipeline
-
-        if self.is_final_transform and other.is_final_transform:
-            raise TransformError(
-                f"Both {self} and {other} are final transforms and cannot be combined."
-            )
-
-        return CompilePipeline(self, other)
-
-    def __mul__(self, n):
-        """Multiply by an integer to create a pipeline with this transform repeated."""
-
-        # Import here to avoid circular import
-        from .compile_pipeline import CompilePipeline  # pylint: disable=import-outside-toplevel
-
-        if not isinstance(n, int):
-            return NotImplemented
-
-        if n < 0:
-            raise ValueError("Cannot multiply transform container by negative integer")
-
-        if self.is_final_transform and n > 1:
-            raise TransformError(
-                f"{self} is a final transform and cannot be applied more than once."
-            )
-
-        return CompilePipeline(self) * n
-
-    __rmul__ = __mul__
-
-
-@Transform.generic_register
-def _apply_to_tape(obj: QuantumScript, transform, *targs, **tkwargs):
-    if transform.tape_transform is None:
-        raise NotImplementedError(
-            f"Transform {transform} has no defined tape implementation, "
-            "and can only be applied when decorating the entire workflow "
-            "with '@qml.qjit' and when it is placed after all transforms "
-            "that only have a tape implementation."
-        )
-    targs, tkwargs = transform.setup_inputs(*targs, **tkwargs)
-    if transform.expand_transform:
-        expanded_tapes, expand_processing = transform.expand_transform(obj, *targs, **tkwargs)
-        transformed_tapes = []
-        processing_and_slices = []
-        start = 0
-        for tape in expanded_tapes:
-            intermediate_tapes, post_processing_fn = transform.tape_transform(
-                tape, *targs, **tkwargs
-            )
-            transformed_tapes.extend(intermediate_tapes)
-            end = start + len(intermediate_tapes)
-            processing_and_slices.append(tuple([post_processing_fn, slice(start, end)]))
-            start = end
-
-        def processing_fn(results):
-            processed_results = [fn(results[slice]) for fn, slice in processing_and_slices]
-            return expand_processing(processed_results)
-
-    else:
-        transformed_tapes, processing_fn = transform.tape_transform(obj, *targs, **tkwargs)
-
-    if transform.is_informative:
-        return processing_fn(transformed_tapes)
-
-    return transformed_tapes, processing_fn
-
-
-def _capture_apply(obj, transform, *targs, **tkwargs):
-    @autograph.wraps(obj)
-    def qfunc_transformed(*args, **kwargs):
-        import jax  # pylint: disable=import-outside-toplevel
-
-        flat_qfunc = capture.flatfn.FlatFn(obj)
-        jaxpr = jax.make_jaxpr(flat_qfunc)(*args, **kwargs)
-        flat_args = jax.tree_util.tree_leaves(args)
-
-        n_args = len(flat_args)
-        n_consts = len(jaxpr.consts)
-        args_slice = slice(0, n_args)
-        consts_slice = slice(n_args, n_args + n_consts)
-        targs_slice = slice(n_args + n_consts, None)
-
-        results = _create_transform_primitive().bind(  # pylint: disable=protected-access
-            *flat_args,
-            *jaxpr.consts,
-            *targs,
-            inner_jaxpr=jaxpr.jaxpr,
-            args_slice=args_slice,
-            consts_slice=consts_slice,
-            targs_slice=targs_slice,
-            tkwargs=tkwargs,
-            transform=transform,
-        )
-
-        assert flat_qfunc.out_tree is not None
-        return jax.tree_util.tree_unflatten(flat_qfunc.out_tree, results)
-
-    return qfunc_transformed
-
-
-@Transform.generic_register
-def apply_to_callable(obj: Callable, transform, *targs, **tkwargs):
-    """Apply a transform to a Callable object."""
-    if obj.__class__.__name__ == "QJIT":
-        raise TransformError(
-            "Functions that are wrapped / decorated with qjit cannot subsequently be"
-            f" transformed with a PennyLane transform (attempted {transform})."
-            f" For the desired affect, ensure that qjit is applied after {transform}."
-        )
-    targs, tkwargs = transform.setup_inputs(*targs, **tkwargs)
-
-    @wraps(obj)
-    def qfunc_transformed(*args, **kwargs):
-        if capture.enabled():
-            return _capture_apply(obj, transform, *targs, **tkwargs)(*args, **kwargs)
-
-        # removes the argument to the qfuncs from the active queuing context.
-        leaves, _ = flatten((args, kwargs), lambda obj: isinstance(obj, Operator))
-        for l in leaves:
-            if isinstance(l, Operator):
-                QueuingManager.remove(l)
-
-        with AnnotatedQueue() as q:
-            qfunc_output = obj(*args, **kwargs)
-
-        tape = QuantumScript.from_queue(q)
-
-        with QueuingManager.stop_recording():
-            if transform.is_informative:
-                transformed_tapes, processing_fn = transform.tape_transform(tape, *targs, **tkwargs)
-            else:
-                transformed_tapes, processing_fn = transform(tape, *targs, **tkwargs)
-
-        if len(transformed_tapes) != 1:
-            raise TransformError(
-                "Impossible to dispatch your transform on quantum function, because more than "
-                "one tape is returned"
-            )
-
-        transformed_tape = transformed_tapes[0]
-
-        if transform.is_informative:
-            return processing_fn(transformed_tapes)
-
-        for op in transformed_tape.operations:
-            apply(op)
-
-        mps = [apply(mp) for mp in transformed_tape.measurements]
-
-        if not mps:
-            return qfunc_output
-
-        if isinstance(qfunc_output, MeasurementProcess):
-            return tuple(mps) if len(mps) > 1 else mps[0]
-
-        if isinstance(qfunc_output, (tuple, list)):
-            return type(qfunc_output)(mps)
-
-        interface = math.get_interface(qfunc_output)
-        return math.asarray(mps, like=interface)
-
-    return qfunc_transformed
-
-
-@Transform.generic_register
-def _apply_to_sequence(obj: Sequence, transform, *targs, **tkwargs):
-    if not all(isinstance(t, QuantumScript) for t in obj):
-        # not a sequence of quantum script, treat as first argument
-        targs, tkwargs = transform.setup_inputs(obj, *targs, **tkwargs)
-        return BoundTransform(transform, args=targs, kwargs=tkwargs)
-    targs, tkwargs = transform.setup_inputs(*targs, **tkwargs)
-    execution_tapes = []
-    batch_fns = []
-    tape_counts = []
-
-    for t in obj:
-        # Preprocess the tapes by applying transforms
-        # to each tape, and storing corresponding tapes
-        # for execution, processing functions, and list of tape lengths.
-        new_tapes, fn = transform(t, *targs, **tkwargs)
-        execution_tapes.extend(new_tapes)
-        batch_fns.append(fn)
-        tape_counts.append(len(new_tapes))
-
-    def processing_fn(res: ResultBatch) -> ResultBatch:
-        """Applies a batch of post-processing functions to results.
-
-        Args:
-            res (ResultBatch): the results of executing a batch of circuits.
-
-        Returns:
-            ResultBatch: results that have undergone classical post processing.
-
-        Closure variables:
-            tape_counts: the number of tapes outputted from each application of the transform.
-            batch_fns: the post processing functions to apply to each sub-batch.
-
-        """
-        count = 0
-        final_results = []
-
-        for f, s in zip(batch_fns, tape_counts):
-            # apply any batch transform post-processing
-            new_res = f(res[count : count + s])
-            final_results.append(new_res)
-            count += s
-
-        return tuple(final_results)
-
-    return tuple(execution_tapes), processing_fn
-
-
-TransformContainer = BoundTransform
 TransformDispatcher = Transform
